+++ Gamefication/src/main/java/net/sf/esfinge/gamification/mechanics/database/nosql/MongoRewardStorage.java
@@ -1,4 +1,4 @@
-package net.sf.esfinge.gamification.mechanics.database.sql;
+package net.sf.esfinge.gamification.mechanics.database.nosql;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -7,23 +7,30 @@ import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.bson.Document;
+
+import com.mongodb.client.MongoCollection;
+
 import net.sf.esfinge.gamification.achievement.Achievement;
 import net.sf.esfinge.gamification.achievement.Reward;
 import net.sf.esfinge.gamification.mechanics.database.Storage;
 
-public class RewardStorage implements Storage {
+public class MongoRewardStorage implements Storage {
 	private Connection connection;
+	private MongoCollection<Document> collection;
 
-	public RewardStorage(Connection c) {
+	public MongoRewardStorage(Connection c) {
 		connection = c;
 	}
 
+	public MongoRewardStorage(MongoCollection<Document> c) {
+		collection = c;
+	}
+
 	public void insert(Object user, Achievement a) throws SQLException {
-		Reward r = (Reward)a;
+		Reward r = (Reward) a;
 		PreparedStatement stmt;
-		stmt = connection
-				.prepareStatement("insert into gamification.reward "
-						+ "(userid, name, used) values (?,?,?)");
+		stmt = connection.prepareStatement("insert into gamification.reward " + "(userid, name, used) values (?,?,?)");
 		stmt.setString(1, user.toString());
 		stmt.setString(2, r.getName());
 		stmt.setBoolean(3, r.isUsed());
@@ -32,9 +39,7 @@ public class RewardStorage implements Storage {
 
 	public Reward select(Object user, String name) throws SQLException {
 		PreparedStatement stmt;
-		stmt = connection
-				.prepareStatement("select * from gamification.reward "
-						+ "where userid=? and name = ?");
+		stmt = connection.prepareStatement("select * from gamification.reward " + "where userid=? and name = ?");
 		stmt.setString(1, user.toString());
 		stmt.setString(2, name);
 		ResultSet rs = stmt.executeQuery();
@@ -45,31 +50,27 @@ public class RewardStorage implements Storage {
 		}
 		return null;
 	}
-	
-	public Map<String, Achievement> select(Object user) throws SQLException{
+
+	public Map<String, Achievement> select(Object user) throws SQLException {
 		Map<String, Achievement> map = new HashMap<String, Achievement>();
 		PreparedStatement stmt;
-		stmt = connection
-				.prepareStatement("select * from gamification.reward "
-						+ "where userid=?");
+		stmt = connection.prepareStatement("select * from gamification.reward " + "where userid=?");
 		stmt.setString(1, user.toString());
 		ResultSet rs = stmt.executeQuery();
-		while(rs.next()){
+		while (rs.next()) {
 			String name = rs.getString("name");
 			boolean u = rs.getBoolean("used");
-			Reward r = new Reward(name,u);
+			Reward r = new Reward(name, u);
 			map.put(r.getName(), r);
 		}
-		
+
 		return map;
 	}
 
 	public void update(Object user, Achievement a) throws SQLException {
-		Reward r = (Reward)a;
+		Reward r = (Reward) a;
 		PreparedStatement stmt;
-		stmt = connection
-				.prepareStatement("update gamification.reward "
-						+ "set used = ? where userid=? and name=?");
+		stmt = connection.prepareStatement("update gamification.reward " + "set used = ? where userid=? and name=?");
 		stmt.setString(2, user.toString());
 		stmt.setString(3, r.getName());
 		stmt.setBoolean(1, r.isUsed());
@@ -79,13 +80,11 @@ public class RewardStorage implements Storage {
 	@Override
 	public void delete(Object user, Achievement p) throws SQLException {
 		PreparedStatement stmt;
-		stmt = connection
-				.prepareStatement("delete from gamification.reward "
-						+ "where userid=? and name = ?");
+		stmt = connection.prepareStatement("delete from gamification.reward " + "where userid=? and name = ?");
 		stmt.setString(1, user.toString());
 		stmt.setString(2, p.getName());
 		stmt.execute();
-		
+
 	}
 
 	@Override
@@ -94,9 +93,9 @@ public class RewardStorage implements Storage {
 		PreparedStatement stmt;
 		stmt = connection.prepareStatement("select userid, name, used from gamification.reward");
 		ResultSet rs = stmt.executeQuery();
-		if(rs != null) {
+		if (rs != null) {
 			map = new HashMap<>();
-			while(rs.next()) {
+			while (rs.next()) {
 				String name = rs.getString("name");
 				boolean used = rs.getBoolean("used");
 				Reward reward = new Reward(name, used);
